requirements.txt
jinja2==2.10
pyhcl==0.3.12
setuptools==40.8.0